revise handling of asynchronous tasks
- remove dependencies to async_tasks and futures_lite
- add on each core a vector with asynchrounous background tasks
