+++ src/arch/x86_64/kernel/core_local.rs
@@ -1,4 +1,5 @@
 use alloc::boxed::Box;
+use alloc::vec::Vec;
 use core::arch::asm;
 use core::cell::Cell;
 use core::ptr;
@@ -10,10 +11,11 @@ use x86_64::VirtAddr;
 
 use super::interrupts::{IrqStatistics, IRQ_COUNTERS};
 use super::CPU_ONLINE;
+use crate::executor::task::AsyncTask;
 use crate::scheduler::{CoreId, PerCoreScheduler};
 
 #[repr(C)]
-pub struct CoreLocal {
+pub(crate) struct CoreLocal {
 	this: *const Self,
 	/// Sequential ID of this CPU Core.
 	core_id: CoreId,
@@ -25,6 +27,8 @@ pub struct CoreLocal {
 	pub kernel_stack: Cell<u64>,
 	/// Interface to the interrupt counters
 	irq_statistics: &'static IrqStatistics,
+	/// Queue of async tasks
+	async_tasks: Cell<Vec<AsyncTask>>,
 }
 
 impl CoreLocal {
@@ -43,6 +47,7 @@ impl CoreLocal {
 			tss: Cell::new(ptr::null_mut()),
 			kernel_stack: Cell::new(0),
 			irq_statistics,
+			async_tasks: Cell::new(Vec::new()),
 		};
 		let this = Box::leak(Box::new(this));
 		this.this = &*this;
@@ -50,6 +55,7 @@ impl CoreLocal {
 		GsBase::write(VirtAddr::from_ptr(this));
 	}
 
+	#[inline]
 	pub fn get() -> &'static Self {
 		debug_assert_ne!(VirtAddr::zero(), GsBase::read());
 		unsafe {
@@ -58,9 +64,19 @@ impl CoreLocal {
 			&*raw
 		}
 	}
+
+	#[inline]
+	pub fn get_mut() -> &'static mut Self {
+		debug_assert_ne!(VirtAddr::zero(), GsBase::read());
+		unsafe {
+			let raw: *mut Self;
+			asm!("mov {}, gs:0", out(reg) raw, options(nomem, nostack, preserves_flags));
+			&mut *raw
+		}
+	}
 }
 
-pub fn core_id() -> CoreId {
+pub(crate) fn core_id() -> CoreId {
 	if cfg!(target_os = "none") {
 		CoreLocal::get().core_id
 	} else {
@@ -68,14 +84,18 @@ pub fn core_id() -> CoreId {
 	}
 }
 
-pub fn core_scheduler() -> &'static mut PerCoreScheduler {
+pub(crate) fn core_scheduler() -> &'static mut PerCoreScheduler {
 	unsafe { &mut *CoreLocal::get().scheduler.get() }
 }
 
-pub fn set_core_scheduler(scheduler: *mut PerCoreScheduler) {
+pub(crate) fn async_tasks() -> &'static mut Vec<AsyncTask> {
+	CoreLocal::get_mut().async_tasks.get_mut()
+}
+
+pub(crate) fn set_core_scheduler(scheduler: *mut PerCoreScheduler) {
 	CoreLocal::get().scheduler.set(scheduler);
 }
 
-pub fn increment_irq_counter(irq_no: u8) {
+pub(crate) fn increment_irq_counter(irq_no: u8) {
 	CoreLocal::get().irq_statistics.inc(irq_no);
 }
