+++ src/executor/mod.rs
@@ -1,211 +1,77 @@
+#![allow(dead_code)]
+
+#[cfg(feature = "tcp")]
 mod device;
-pub(crate) mod runtime;
+#[cfg(feature = "tcp")]
+pub(crate) mod network;
+pub(crate) mod task;
 
-use alloc::boxed::Box;
-use core::ops::DerefMut;
-use core::sync::atomic::{AtomicU16, Ordering};
-use core::task::Poll;
+use alloc::sync::Arc;
+use alloc::task::Wake;
+use core::future::Future;
+use core::sync::atomic::{AtomicBool, Ordering};
+use core::task::{Context, Poll, Waker};
 
-use futures_lite::future;
-use hermit_sync::InterruptTicketMutex;
-use smoltcp::iface::{SocketHandle, SocketSet};
-#[cfg(feature = "dhcpv4")]
-use smoltcp::socket::dhcpv4;
-use smoltcp::socket::{tcp, udp, AnySocket};
-use smoltcp::time::{Duration, Instant};
-#[cfg(feature = "dhcpv4")]
-use smoltcp::wire::{IpCidr, Ipv4Address, Ipv4Cidr};
+use hermit_sync::without_interrupts;
 
-use crate::arch;
-use crate::executor::device::HermitNet;
-use crate::executor::runtime::spawn;
+use crate::arch::core_local::*;
+use crate::executor::task::AsyncTask;
+use crate::scheduler::task::TaskHandle;
 
-pub(crate) enum NetworkState<'a> {
-	Missing,
-	InitializationFailed,
-	Initialized(Box<NetworkInterface<'a>>),
+struct TaskNotify {
+	/// The single task executor .
+	handle: TaskHandle,
+	/// A flag to ensure a wakeup is not "forgotten" before the next `block_current_task`
+	unparked: AtomicBool,
 }
 
-impl<'a> NetworkState<'a> {
-	pub fn as_nic_mut(&mut self) -> Result<&mut NetworkInterface<'a>, &'static str> {
-		match self {
-			NetworkState::Initialized(nic) => Ok(nic),
-			_ => Err("Network is not initialized!"),
+impl TaskNotify {
+	pub fn new() -> Self {
+		Self {
+			handle: core_scheduler().get_current_task_handle(),
+			unparked: AtomicBool::new(false),
 		}
 	}
 }
 
-pub(crate) type Handle = SocketHandle;
-
-static LOCAL_ENDPOINT: AtomicU16 = AtomicU16::new(0);
-pub(crate) static NIC: InterruptTicketMutex<NetworkState<'_>> =
-	InterruptTicketMutex::new(NetworkState::Missing);
-
-pub(crate) struct NetworkInterface<'a> {
-	iface: smoltcp::iface::Interface,
-	sockets: SocketSet<'a>,
-	device: HermitNet,
-	#[cfg(feature = "dhcpv4")]
-	dhcp_handle: SocketHandle,
-}
-
-#[cfg(target_arch = "x86_64")]
-fn start_endpoint() -> u16 {
-	((unsafe { core::arch::x86_64::_rdtsc() }) % (u16::MAX as u64))
-		.try_into()
-		.unwrap()
-}
-
-#[cfg(target_arch = "aarch64")]
-fn start_endpoint() -> u16 {
-	use core::arch::asm;
-	let value: u64;
-
-	unsafe {
-		asm!(
-			"mrs {value}, cntpct_el0",
-			value = out(reg) value,
-			options(nostack),
-		);
+impl Wake for TaskNotify {
+	fn wake(self: Arc<Self>) {
+		self.wake_by_ref()
 	}
 
-	(value % (u16::MAX as u64)).try_into().unwrap()
-}
-
-#[inline]
-pub(crate) fn now() -> Instant {
-	let microseconds = arch::processor::get_timer_ticks() + arch::get_boot_time();
-	Instant::from_micros_const(microseconds.try_into().unwrap())
-}
-
-async fn network_run() {
-	future::poll_fn(|cx| match NIC.lock().deref_mut() {
-		NetworkState::Initialized(nic) => {
-			nic.poll_common(now());
-
-			// this background task will never stop
-			// => wakeup ourself
-			cx.waker().clone().wake();
-
-			Poll::Pending
+	fn wake_by_ref(self: &Arc<Self>) {
+		// Make sure the wakeup is remembered until the next `park()`.
+		let unparked = self.unparked.swap(true, Ordering::AcqRel);
+		if !unparked {
+			trace!("Waker wakes async task {}", self.handle.get_id());
+			core_scheduler().custom_wakeup(self.handle);
 		}
-		_ => Poll::Ready(()),
-	})
-	.await
-}
-
-pub(crate) fn init() {
-	info!("Try to initialize network!");
-
-	// initialize variable, which contains the next local endpoint
-	LOCAL_ENDPOINT.store(start_endpoint(), Ordering::Relaxed);
-
-	let mut guard = NIC.lock();
-
-	*guard = NetworkInterface::create();
-
-	if let NetworkState::Initialized(nic) = guard.deref_mut() {
-		let time = now();
-		nic.poll_common(time);
-		let wakeup_time = nic
-			.poll_delay(time)
-			.map(|d| crate::arch::processor::get_timer_ticks() + d.total_micros());
-		crate::core_scheduler().add_network_timer(wakeup_time);
-
-		spawn(network_run()).detach();
 	}
 }
 
-impl<'a> NetworkInterface<'a> {
-	pub(crate) fn create_udp_handle(&mut self) -> Result<Handle, ()> {
-		// Must fit mDNS payload of at least one packet
-		let udp_rx_buffer =
-			udp::PacketBuffer::new(vec![udp::PacketMetadata::EMPTY; 4], vec![0; 1024]);
-		// Will not send mDNS
-		let udp_tx_buffer = udp::PacketBuffer::new(vec![udp::PacketMetadata::EMPTY], vec![0; 0]);
-		let udp_socket = udp::Socket::new(udp_rx_buffer, udp_tx_buffer);
-		let udp_handle = self.sockets.add(udp_socket);
+pub(crate) fn run() {
+	let waker = Waker::noop();
+	let mut cx = Context::from_waker(&waker);
 
-		Ok(udp_handle)
-	}
+	async_tasks().retain_mut(|task| {
+		trace!("Run async task {}", task.id());
 
-	pub(crate) fn create_tcp_handle(&mut self) -> Result<Handle, ()> {
-		let tcp_rx_buffer = tcp::SocketBuffer::new(vec![0; 65535]);
-		let tcp_tx_buffer = tcp::SocketBuffer::new(vec![0; 65535]);
-		let mut tcp_socket = tcp::Socket::new(tcp_rx_buffer, tcp_tx_buffer);
-		tcp_socket.set_nagle_enabled(true);
-		let tcp_handle = self.sockets.add(tcp_socket);
+		match task.poll(&mut cx) {
+			Poll::Ready(()) => false,
+			Poll::Pending => true,
+		}
+	});
+}
 
-		Ok(tcp_handle)
-	}
+/// Spawns a future on the executor.
+pub(crate) fn spawn<F>(future: F)
+where
+	F: Future<Output = ()> + Send + 'static,
+{
+	without_interrupts(|| async_tasks().push(AsyncTask::new(future)));
+}
 
-	pub(crate) fn poll_common(&mut self, timestamp: Instant) {
-		let _ = self
-			.iface
-			.poll(timestamp, &mut self.device, &mut self.sockets);
-
-		#[cfg(feature = "dhcpv4")]
-		match self
-			.sockets
-			.get_mut::<dhcpv4::Socket<'_>>(self.dhcp_handle)
-			.poll()
-		{
-			None => {}
-			Some(dhcpv4::Event::Configured(config)) => {
-				info!("DHCP config acquired!");
-				info!("IP address:      {}", config.address);
-				self.iface.update_ip_addrs(|addrs| {
-					if let Some(dest) = addrs.iter_mut().next() {
-						*dest = IpCidr::Ipv4(config.address);
-					} else if addrs.push(IpCidr::Ipv4(config.address)).is_err() {
-						info!("Unable to update IP address");
-					}
-				});
-				if let Some(router) = config.router {
-					info!("Default gateway: {}", router);
-					self.iface
-						.routes_mut()
-						.add_default_ipv4_route(router)
-						.unwrap();
-				} else {
-					info!("Default gateway: None");
-					self.iface.routes_mut().remove_default_ipv4_route();
-				}
-
-				for (i, s) in config.dns_servers.iter().enumerate() {
-					info!("DNS server {}:    {}", i, s);
-				}
-			}
-			Some(dhcpv4::Event::Deconfigured) => {
-				info!("DHCP lost config!");
-				let cidr = Ipv4Cidr::new(Ipv4Address::UNSPECIFIED, 0);
-				self.iface.update_ip_addrs(|addrs| {
-					if let Some(dest) = addrs.iter_mut().next() {
-						*dest = IpCidr::Ipv4(cidr);
-					}
-				});
-				self.iface.routes_mut().remove_default_ipv4_route();
-			}
-		};
-	}
-
-	pub(crate) fn poll_delay(&mut self, timestamp: Instant) -> Option<Duration> {
-		self.iface.poll_delay(timestamp, &self.sockets)
-	}
-
-	#[allow(dead_code)]
-	pub(crate) fn get_socket<T: AnySocket<'a>>(&self, handle: SocketHandle) -> &T {
-		self.sockets.get(handle)
-	}
-
-	pub(crate) fn get_mut_socket<T: AnySocket<'a>>(&mut self, handle: SocketHandle) -> &mut T {
-		self.sockets.get_mut(handle)
-	}
-
-	pub(crate) fn get_socket_and_context<T: AnySocket<'a>>(
-		&mut self,
-		handle: SocketHandle,
-	) -> (&mut T, &mut smoltcp::iface::Context) {
-		(self.sockets.get_mut(handle), self.iface.context())
-	}
+pub fn init() {
+	#[cfg(all(feature = "tcp", not(feature = "newlib")))]
+	crate::executor::network::init();
 }
