+++ src/fd/socket/mod.rs
@@ -4,7 +4,7 @@ use core::ops::DerefMut;
 use core::sync::atomic::Ordering;
 
 use crate::errno::*;
-use crate::executor::{NetworkState, NIC};
+use crate::executor::network::{NetworkState, NIC};
 use crate::fd::{get_object, insert_object, FD_COUNTER, OBJECT_MAP};
 use crate::syscalls::net::*;
