+++ src/scheduler/mod.rs
@@ -4,8 +4,6 @@ use alloc::rc::Rc;
 #[cfg(feature = "smp")]
 use alloc::vec::Vec;
 use core::cell::RefCell;
-#[cfg(feature = "tcp")]
-use core::ops::DerefMut;
 use core::sync::atomic::{AtomicU32, Ordering};
 
 use crossbeam_utils::Backoff;
@@ -75,9 +73,6 @@ pub struct PerCoreScheduler {
 	finished_tasks: VecDeque<Rc<RefCell<Task>>>,
 	/// Queue of blocked tasks, sorted by wakeup time.
 	blocked_tasks: BlockedTaskQueue,
-	/// Queue of blocked tasks, sorted by wakeup time.
-	#[cfg(feature = "tcp")]
-	blocked_async_tasks: VecDeque<TaskHandle>,
 	/// Queues to handle incoming requests from the other cores
 	#[cfg(feature = "smp")]
 	input: InterruptTicketMutex<SchedulerInput>,
@@ -301,9 +296,8 @@ impl PerCoreScheduler {
 	#[inline]
 	pub fn handle_waiting_tasks(&mut self) {
 		without_interrupts(|| {
-			#[cfg(feature = "tcp")]
-			self.wakeup_async_tasks();
-			self.blocked_tasks.handle_waiting_tasks()
+			crate::executor::run();
+			self.blocked_tasks.handle_waiting_tasks();
 		});
 	}
 
@@ -338,42 +332,6 @@ impl PerCoreScheduler {
 	#[inline]
 	pub fn add_network_timer(&mut self, wakeup_time: Option<u64>) {
 		without_interrupts(|| self.blocked_tasks.add_network_timer(wakeup_time))
-	}
-
-	#[cfg(feature = "tcp")]
-	#[inline]
-	pub fn block_current_async_task(&mut self) {
-		without_interrupts(|| {
-			self.blocked_async_tasks
-				.push_back(self.get_current_task_handle());
-			self.blocked_tasks.add(self.current_task.clone(), None)
-		});
-	}
-
-	#[cfg(feature = "tcp")]
-	#[inline]
-	pub fn wakeup_async_tasks(&mut self) {
-		let mut has_tasks = false;
-
-		without_interrupts(|| {
-			while let Some(task) = self.blocked_async_tasks.pop_front() {
-				has_tasks = true;
-				self.custom_wakeup(task)
-			}
-
-			if !has_tasks {
-				if let Some(mut guard) = crate::executor::NIC.try_lock() {
-					if let crate::executor::NetworkState::Initialized(nic) = guard.deref_mut() {
-						let time = crate::executor::now();
-						nic.poll_common(time);
-						let wakeup_time = nic
-							.poll_delay(time)
-							.map(|d| crate::arch::processor::get_timer_ticks() + d.total_micros());
-						self.add_network_timer(wakeup_time);
-					}
-				}
-			}
-		});
 	}
 
 	#[inline]
@@ -564,6 +522,10 @@ impl PerCoreScheduler {
 
 		loop {
 			interrupts::disable();
+
+			// run async tasks
+			crate::executor::run();
+
 			// do housekeeping
 			self.cleanup_tasks();
 
@@ -725,8 +687,6 @@ pub fn add_current_core() {
 		ready_queue: PriorityTaskQueue::new(),
 		finished_tasks: VecDeque::new(),
 		blocked_tasks: BlockedTaskQueue::new(),
-		#[cfg(feature = "tcp")]
-		blocked_async_tasks: VecDeque::new(),
 		#[cfg(feature = "smp")]
 		input: InterruptTicketMutex::new(SchedulerInput::new()),
 	});
